Description:modify dlp review problem
Match-id-2302e5c3b0a9daa086c7b4092687ba4ccd414b77

+++ interfaces/inner_api/dlp_fuse/include/dlp_link_file.h
@@ -54,7 +54,7 @@ public:
         return dlpLinkName_;
     }
 
-    int32_t Truncate(off_t modifySize);
+    int32_t Truncate(uint32_t modifySize);
 
 private:
     std::string dlpLinkName_;
